base64 decode arg casts
```
src/knownhost.c:173:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/knownhost.c:173:42: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:173:42: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:173:42: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:173:42: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:180:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/knownhost.c:180:42: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:180:42: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:180:42: warning: implicit conversion loses integer precision: 'unsigned long long' to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:180:42: warning: implicit conversion loses integer precision: 'unsigned long' to 'unsigned int' [-Wshorten-64-to-32]
src/pem.c:394:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/pem.c:394:39: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/pem.c:394:39: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/pem.c:394:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'unsigned int' [-Wshorten-64-to-32]
src/pem.c:394:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:614:..: warning C4244: 'function': conversion from '__int64' to 'unsigned int', possible loss of data
src/userauth.c:614:49: warning: conversion from 'long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:614:49: warning: conversion from 'long long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:614:49: warning: implicit conversion loses integer precision: 'long long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:614:49: warning: implicit conversion loses integer precision: 'long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:718:..: warning C4244: 'function': conversion from '__int64' to 'unsigned int', possible loss of data
src/userauth.c:718:49: warning: conversion from 'long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:718:49: warning: conversion from 'long long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:718:49: warning: implicit conversion loses integer precision: 'long long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:718:49: warning: implicit conversion loses integer precision: 'long' to 'unsigned int' [-Wshorten-64-to-32]
```

+++ src/userauth.c
@@ -611,7 +611,7 @@ memory_read_publickey(LIBSSH2_SESSION * session, unsigned char **method,
     }
 
     if(libssh2_base64_decode(session, (char **) &tmp, &tmp_len,
-                              (char *) sp1, sp2 - sp1)) {
+                              (char *) sp1, (unsigned int)(sp2 - sp1))) {
         LIBSSH2_FREE(session, pubkey);
         return _libssh2_error(session, LIBSSH2_ERROR_FILE,
                                   "Invalid key data, not base64 encoded");
@@ -715,7 +715,7 @@ file_read_publickey(LIBSSH2_SESSION * session, unsigned char **method,
     }
 
     if(libssh2_base64_decode(session, (char **) &tmp, &tmp_len,
-                              (char *) sp1, sp2 - sp1)) {
+                              (char *) sp1, (unsigned int)(sp2 - sp1))) {
         LIBSSH2_FREE(session, pubkey);
         return _libssh2_error(session, LIBSSH2_ERROR_FILE,
                               "Invalid key data, not base64 encoded");
